+++ src/main/java/com/github/mauricioaniche/ck/metric/NOperands.java
@@ -10,6 +10,7 @@ import org.eclipse.jdt.core.dom.SimpleName;
 
 import com.github.mauricioaniche.ck.CKNumber;
 import com.github.mauricioaniche.ck.CKReport;
+import com.github.mauricioaniche.ck.metric.dependent.NUniqueOperands;
 
 /**
  * Number of operands
