+++ src/main/java/com/github/mauricioaniche/ck/metric/NOperators.java
@@ -11,6 +11,7 @@ import org.eclipse.jdt.core.dom.PrefixExpression;
 
 import com.github.mauricioaniche.ck.CKNumber;
 import com.github.mauricioaniche.ck.CKReport;
+import com.github.mauricioaniche.ck.metric.dependent.NUniqueOperators;
 
 /**
  * @author klerisson
