+++ src/main/java/com/github/mauricioaniche/ck/metric/dependent/NUniqueOperands.java
@@ -1,4 +1,4 @@
-package com.github.mauricioaniche.ck.metric;
+package com.github.mauricioaniche.ck.metric.dependent;
 
 import java.util.HashSet;
 import java.util.Set;
